Minor change - int to size_t to get ride of warnings
git-svn-id: https://svn.rice.edu/r/software_0/ObjectQueryModel@2654 522aa1da-077c-40de-a48b-1771163b7009

+++ src/builtInPDBObjects/headers/CheckEmployees.h
@@ -15,7 +15,6 @@
  *  limitations under the License.                                           *
  *                                                                           *
  *****************************************************************************/
-
 #ifndef CHECK_EMPLOYEES_H
 #define CHECK_EMPLOYEES_H
 
@@ -33,7 +32,7 @@ namespace pdb {
 // if he doesn't manage anyone other than "nameToExclude".  If he does manage
 // someone other than "nameToExclude", then the query will return all of the
 // employees for the supervisor that don't have that name 
-class CheckEmployee : public Selection <Vector <Handle <Employee>>, Supervisor> {
+class CheckEmployee: public Selection<Vector<Handle<Employee>>, Supervisor> {
 
 private:
 
@@ -43,16 +42,17 @@ public:
 
 	ENABLE_DEEP_COPY
 
-	CheckEmployee () {}
+	CheckEmployee() {
+	}
 
-	CheckEmployee (std :: string &nameToExcludeIn) {
+	CheckEmployee(std::string &nameToExcludeIn) {
 		nameToExclude = nameToExcludeIn;
 	}
-	
-	SimpleLambda <bool> getSelection (Handle <Supervisor> &checkMe) override {
-		return makeLambda (checkMe, [&] () -> bool {
+
+	SimpleLambda<bool> getSelection(Handle<Supervisor> &checkMe) override {
+		return makeLambda(checkMe, [&] () -> bool {
 			int numEmployees = checkMe->getNumEmployees ();
-			for (int i = 0; i < numEmployees; i++) {
+			for (unsigned int i = 0; i < numEmployees; i++) {
 				if (*(checkMe->getEmp (i)->getName ()) != nameToExclude) {
 					return true;
 				}
@@ -61,33 +61,29 @@ public:
 		});
 	}
 
-
-
-        SimpleLambda <bool> getProjectionSelection (Handle <Vector<Handle<Employee>>> &checkMe) override {
-                return makeLambda (checkMe, [&] () -> bool {
-                        int numEmployees = checkMe->size();
-                        //std :: cout << "numEmployees=" << numEmployees << std :: endl;
-                        for (int i = 0; i < numEmployees; i++) {
-                                //std :: cout << i << ":"  << *(*checkMe)[i]->getName () << std :: endl;
-                                if ((*(*checkMe)[i]->getName ()) != nameToExclude) {
-                                        return true;
-                                }
-                        }
-                        return false;
-                });
-        }
-
-
+	SimpleLambda<bool> getProjectionSelection(Handle<Vector<Handle<Employee>>>&checkMe) override {
+		return makeLambda (checkMe, [&] () -> bool {
+			size_t numEmployees = checkMe->size();
+			//std :: cout << "numEmployees=" << numEmployees << std :: endl;
+			for (size_t i = 0; i < numEmployees; i++) {
+				//std :: cout << i << ":"  << *(*checkMe)[i]->getName () << std :: endl;
+				if ((*(*checkMe)[i]->getName ()) != nameToExclude) {
+					return true;
+				}
+			}
+			return false;
+		});
+	}
 
 	SimpleLambda <Handle <Vector <Handle <Employee>>>> getProjection (Handle <Supervisor> &checkMe) override {
 		return makeLambda (checkMe, [&] {
 			Handle <Vector <Handle <Employee>>> returnVal = makeObject <Vector <Handle <Employee>>> (10);
-			int numEmployees = checkMe->getNumEmployees ();
-                        //std :: cout << "numEmployees=" << numEmployees << std :: endl;
-			for (int i = 0; i < numEmployees; i++) {
-                                //std :: cout << i << ":"  << *(checkMe->getEmp (i)->getName ()) << std :: endl;
+			size_t numEmployees = checkMe->getNumEmployees ();
+			//std :: cout << "numEmployees=" << numEmployees << std :: endl;
+			for (size_t i = 0; i < numEmployees; i++) {
+				//std :: cout << i << ":"  << *(checkMe->getEmp (i)->getName ()) << std :: endl;
 				if (*(checkMe->getEmp (i)->getName ()) != nameToExclude) {
-					returnVal->push_back (checkMe->getEmp (i));	
+					returnVal->push_back (checkMe->getEmp (i));
 				}
 			}
